src/services/secure_computation_service.py
"""
PySyft Secure Computation Service for ClimateCoach
Enables secure and private data handling
"""

import syft as sy
import os
import logging

logger = logging.getLogger(__name__)

class SecureComputationService:
    """
    Service to enable secure data computation using PySyft
    """
    
    def __init__(self):
        """Initialize PySyft Virtual Machine"""
        self.vm = sy.VirtualMachine(name="climatecoach_vm")
        self.root_connection = self.vm.get_root_client()
        logger.info("Secure Computation Service initialized with PySyft VM")
        
    def encrypt_data(self, data):
        """Encrypt data using PySyft"""
        try:
            encrypted_data = data.send(self.root_connection)
            logger.info("Data encrypted successfully")
            return encrypted_data
        except Exception as e:
            logger.error(f"Error encrypting data: {e}")
            return None
        
    def decrypt_data(self, encrypted_data):
        """Decrypt data using PySyft"""
        try:
            decrypted_data = encrypted_data.get()
            logger.info("Data decrypted successfully")
            return decrypted_data
        except Exception as e:
            logger.error(f"Error decrypting data: {e}")
            return None
    
    def perform_secure_computation(self, encrypted_data, computation_function):
        """Perform encrypted computation"""
        try:
            # Perform computation
            result = computation_function(encrypted_data)
            logger.info("Secure computation executed successfully")
            return result.get()
        except Exception as e:
            logger.error(f"Error performing secure computation: {e}")
            return None

if __name__ == "__main__":
    # Example usage
    secure_service = SecureComputationService()
    
    # Example data
    example_data = sy.randn(1, 10)
    encrypted_data = secure_service.encrypt_data(example_data)
    decrypted_data = secure_service.decrypt_data(encrypted_data)
    print("Decrypted Data:", decrypted_data)

