requirements.txt
# Core Framework & Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Frontend UI
streamlit==1.28.2
streamlit-option-menu==0.3.6
streamlit-plotly-events==0.0.6
streamlit-aggrid==0.3.4

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# AI & ML for Carbon Estimation & Recommendations 
openai==1.3.8
langchain==0.0.348
langchain-openai==0.0.2
crewai==0.1.55
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
tslearn==0.6.3
prophet==1.1.4
transformers==4.36.0
torch==2.1.0

# Data Visualization & Analytics
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
altair==5.2.0
folium==0.15.1
streamlit-folium==0.15.0

# Data Processing & APIs
requests==2.31.0
aiohttp==3.9.1
pillow==10.1.0
geopy==2.4.1
pyproj==3.6.1

# AWS Cloud Services
boto3==1.34.0
botocore==1.34.0

# Carbon Footprint & Environmental Data
meteostat==1.6.5
carbonfootprint==1.0.3
environmental-apis==0.2.1

# Time Series & Habit Detection
tslearn==0.6.3
prophet==1.1.4
scipy==1.11.4
sklearn==1.3.2

# Gamification & Behavioral Analytics
mixpanel==4.10.0
posthog==3.1.0

# Privacy  Offline-First
tensorflow-lite==2.14.0
onnxruntime==1.16.3
pysyft==0.8.7
transformers==4.36.0
roberta==1.0.0
tslearn==0.6.3
prophet==1.1.4

# Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3
psutil==5.9.6

# AI Offline
openai-offline==1.3.8

# Utilities & Configuration
python-dotenv==1.0.0
celery==5.3.4
click==8.1.7
pyyaml==6.0.1
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# GDPR Compliance & Privacy
python-gdpr==0.1.0
data-anonymization==1.0.0
